SEQNG-58 Remove legacy seqexec events and rename extisting SharedModel package #time 2h
