+++ modules/edu.gemini.seqexec.web/edu.gemini.seqexec.web.client/src/test/scala/edu/gemini/seqexec/web/client/model/SeqexecCircuitSequenceDisplayHandlerSpec.scala
@@ -2,7 +2,7 @@ package edu.gemini.seqexec.web.client.model
 
 import diode.ActionResult.ModelUpdate
 import diode.RootModelRW
-import edu.gemini.seqexec.model.SharedModel.SequenceView
+import edu.gemini.seqexec.model.Model.SequenceView
 import org.scalatest.prop.PropertyChecks
 import org.scalatest.{FlatSpec, Matchers}
